refactor: separate nav items into new file
