+++ src/components/types/index.ts
@@ -1,0 +1,1 @@
+export { navItems } from "./nav-items.type";
