+++ src/components/types/nav-items.type.ts
@@ -1,0 +1,27 @@
+type NavItem = {
+  name: string;
+  href: string;
+};
+
+export const navItems: NavItem[] = [
+  {
+    name: "About",
+    href: "#",
+  },
+  {
+    name: "Careers",
+    href: "#",
+  },
+  {
+    name: "Events",
+    href: "#",
+  },
+  {
+    name: "Products",
+    href: "#",
+  },
+  {
+    name: "Support",
+    href: "#",
+  },
+];
